refactor :: 이상한 심볼릭 레퍼런스 수정
