+++ src/main/java/com/example/bssm_dev/domain/user/service/UserLoginService.java
@@ -29,14 +29,4 @@ public class UserLoginService {
         return userLoginResponse;
     }
 
-    public boolean isUserExists(String email) {
-        return userRepository.existsByEmail(email);
-    }
-
-    public UserLoginResponse getUserByEmail(String email) {
-        User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new RuntimeException("User not found"));
-
-        return userMapper.toUserLoginResponse(user);
-    }
 }
